agent/pyproject.toml
[project]
name = "agentcore-agent"
version = "0.1.0"
description = "Local AgentCore Runtime with Strands Agent"
requires-python = ">=3.12"
dependencies = [
    "strands-agents[ollama]>=1.0.0",
    "strands-agents-tools>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "aiosqlite>=0.20.0",
    "ag-ui-protocol>=0.1.5",
    "ag_ui_strands~=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["main.py"]

[tool.hatch.build.targets.wheel.force-include]
"main.py" = "main.py"
